Revert "Features/genres merged with main"
