+++ Controllers/GenresController.cs
@@ -1,9 +1,6 @@
 ﻿using Bookstore.Data;
 using Bookstore.Models;
-using Bookstore.Models.ViewModels;
-using Bookstore.Services;
 using Microsoft.AspNetCore.Mvc;
-using System.Diagnostics;
 
 namespace Bookstore.Controllers
 {
@@ -16,53 +13,14 @@ namespace Bookstore.Controllers
             _service = service;
         }
 
-        public async Task<IActionResult> Index()
+        public IActionResult Index()
         {
-            List<Genre> genres = await _service.FindAllAsync();
-            return View(genres);
+            return View(_service.FindAll());
         }
 
         public IActionResult Create()
         {
             return View();
-        }
-
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create(Genre genre)
-        {
-            if(!ModelState.IsValid)
-            {
-                return View();
-            }
-
-            await _service.InsertAsync(genre);
-
-            return RedirectToAction(nameof(Index));
-        }
-
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id is null)
-            {
-                return RedirectToAction(nameof(Error), new { message = "Id não fornecido" });
-            } 
-            var obj = await _service.FindByIdAsync(id.Value);
-            if (obj is null)
-            {
-                return RedirectToAction(nameof(Error), new { message = "Id não encontrado" });
-            }
-            return View(obj);
-        }
-
-        public IActionResult Error(string? message)
-        {
-            var viewModel = new ErrorViewModel
-            {
-                Message = message,
-                RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
-            };
-            return View(viewModel);
         }
     }
 }
