+++ Models/ViewModels/ErrorViewModel.cs
@@ -4,8 +4,6 @@ namespace Bookstore.Models.ViewModels
     {
         public string? RequestId { get; set; }
 
-        public string? Message { get; set; }
-
         public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
     }
 }
