+++ Views/Genres/Create.cshtml
@@ -1,12 +1,12 @@
 ﻿@model Bookstore.Models.Genre
 
 @{
-    ViewData["Title"] = "Adicionar Gênero Literário";
+    ViewData["Title"] = "Novo Gênero Literário";
 }
 
 <h1>@ViewData["Title"]</h1>
 
-<h4 class="lead">Preencha os campos abaixo</h4>
+<h4>Genre</h4>
 <hr />
 <div class="row">
     <div class="col-md-4">
@@ -17,7 +17,7 @@
                 <input asp-for="Name" class="form-control" />
                 <span asp-validation-for="Name" class="text-danger"></span>
             </div>
-            <div class="form-group my-2">
+            <div class="form-group">
                 <input type="submit" value="Create" class="btn btn-primary" />
             </div>
         </form>
@@ -25,7 +25,7 @@
 </div>
 
 <div>
-    <a asp-action="Index">Voltar para a lista</a>
+    <a asp-action="Index">Back to List</a>
 </div>
 
 @section Scripts {
