--- Views/Genres/Delete.cshtml
@@ -1,33 +1,0 @@
-﻿@model Bookstore.Models.Genre
-
-@{
-    ViewData["Title"] = "Delete";
-}
-
-<h1>Deletar</h1>
-
-<h3>Voce tem certeza que deseja apagar isto?</h3>
-<div>
-    <h4>Gêneros literários</h4>
-    <hr />
-    <dl class="row">
-        <dt class = "col-sm-2">
-            @Html.DisplayNameFor(model => model.Id)
-        </dt>
-        <dd class = "col-sm-10">
-            @Html.DisplayFor(model => model.Id)
-        </dd>
-        <dt class="col-sm-2">
-            @Html.DisplayNameFor(model => model.Name)
-        </dt>
-        <dd class="col-sm-10">
-            @Html.DisplayFor(model => model.Name)
-        </dd>
-    </dl>
-    
-    <form asp-action="Delete">
-        <input type="hidden" asp-for="Id" />
-        <input type="submit" value="Apagar" class="btn btn-danger" /> |
-        <a asp-action="Index">Voltar aos Gêneros Literários</a>
-    </form>
-</div>
