+++ Views/Shared/Error.cshtml
@@ -4,9 +4,8 @@
     ViewData["Title"] = "Error";
 }
 
-<h1 class="text-danger">Erro</h1>
-<h2 class="text-danger">Um erro ocorreu ao tentar fazer a requisição</h2>
-<h3 class="text-danger">@Model.Message</h3>
+<h1 class="text-danger">Error.</h1>
+<h2 class="text-danger">An error occurred while processing your request.</h2>
 
 @if (Model.ShowRequestId)
 {
@@ -14,3 +13,14 @@
         <strong>Request ID:</strong> <code>@Model.RequestId</code>
     </p>
 }
+
+<h3>Development Mode</h3>
+<p>
+    Swapping to <strong>Development</strong> environment will display more detailed information about the error that occurred.
+</p>
+<p>
+    <strong>The Development environment shouldn't be enabled for deployed applications.</strong>
+    It can result in displaying sensitive information from exceptions to end users.
+    For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
+    and restarting the app.
+</p>
